Consolidate more services
