--- src/undo-log-setup.ts
@@ -1,7 +1,0 @@
-export interface UndoLogSetup {
-  install(): Promise<string[]>;
-  uninstall(): Promise<void>;
-  enableDebugMode(enabled: boolean): Promise<void>;
-  addTable(name: string, channel: number): Promise<void>;
-  removeTable(name: string): Promise<void>;
-}
